requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.16.0
numpy==2.4.2
openai==2.17.0
packaging==26.0
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==12.1.0
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
qdrant-client==1.16.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.54.0
tenacity==9.1.3
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.5.3
